backend/requirements.txt
asgiref==3.8.1; python_version >= '3.8'
bcrypt==4.2.1; python_version >= '3.7'
django==5.1.3; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-filter==24.3; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
drf-yasg==1.21.8; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
markdown==3.7; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pillow==11.0.0; python_version >= '3.9'
pyjwt==2.10.0; python_version >= '3.9'
python-decouple==3.8
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
sqlparse==0.5.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'